+++ app/src/main/java/eu/faircode/email/FragmentDialogDataSaver.java
@@ -24,7 +24,9 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.net.Uri;
 import android.os.Bundle;
+import android.provider.Settings;
 import android.text.SpannableStringBuilder;
 import android.text.style.RelativeSizeSpan;
 
@@ -33,36 +35,43 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.preference.PreferenceManager;
 
-public class FragmentDialogNotifications extends FragmentDialogBase {
+public class FragmentDialogDataSaver extends FragmentDialogBase {
     @NonNull
     @Override
     public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
         final Context context = getContext();
 
         SpannableStringBuilder ssb = new SpannableStringBuilderEx();
-        ssb.append(context.getString(R.string.title_hint_notifications_remark));
+        ssb.append(context.getString(R.string.title_hint_data_saver));
         ssb.append("\n\n");
         int start = ssb.length();
         ssb.append(context.getString(R.string.title_hint_dismiss));
         ssb.setSpan(new RelativeSizeSpan(HtmlHelper.FONT_SMALL), start, ssb.length(), 0);
 
+        Intent saver = new Intent(
+                Settings.ACTION_IGNORE_BACKGROUND_DATA_RESTRICTIONS_SETTINGS,
+                Uri.parse("package:" + BuildConfig.APPLICATION_ID));
+
+        Intent settings = (saver.resolveActivity(context.getPackageManager()) == null
+                ? new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS,
+                Uri.parse("package:" + BuildConfig.APPLICATION_ID))
+                : saver);
+
         return new AlertDialog.Builder(context)
-                .setIcon(R.drawable.twotone_notifications_off_24)
-                .setTitle(R.string.title_hint_notifications)
+                .setIcon(R.drawable.twotone_data_saver_off_24)
+                .setTitle(R.string.title_setup_data)
                 .setMessage(ssb)
                 .setPositiveButton(R.string.title_fix, new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        Intent intent = new Intent(context, ActivitySetup.class)
-                                .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                        context.startActivity(intent);
+                        context.startActivity(settings);
                     }
                 })
                 .setNegativeButton(R.string.title_dismiss, new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
                         SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
-                        prefs.edit().putBoolean("notifications_reminder", false).apply();
+                        prefs.edit().putBoolean("datasaver_reminder", false).apply();
                     }
                 })
                 .create();