+++ app/src/main/java/eu/faircode/email/FragmentOptionsMisc.java
@@ -216,6 +216,7 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
     private SwitchCompat swWebp;
     private SwitchCompat swAnimate;
     private SwitchCompat swEasyCorrect;
+    private SwitchCompat swPastePlain;
     private SwitchCompat swInfra;
     private SwitchCompat swTldFlags;
     private SwitchCompat swJsonLd;
@@ -288,7 +289,7 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
             "exact_alarms",
             "native_dkim", "native_arc", "native_arc_whitelist",
             "webp", "animate_images",
-            "easy_correct", "infra", "tld_flags", "json_ld", "dup_msgids", "thread_byref", "save_user_flags", "mdn",
+            "easy_correct", "paste_plain", "infra", "tld_flags", "json_ld", "dup_msgids", "thread_byref", "save_user_flags", "mdn",
             "app_chooser", "app_chooser_share", "adjacent_links", "adjacent_documents", "adjacent_portrait", "adjacent_landscape",
             "delete_confirmation", "global_keywords", "test_iab"
     ));
@@ -308,6 +309,7 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
             "raw_asked", "all_read_asked", "delete_asked",
             "cc_bcc", "inline_image_hint", "compose_reference", "send_dialog",
             "setup_reminder", "was_ignoring", "setup_advanced",
+            "notifications_reminder", "datasaver_reminder",
             "signature_images_hint",
             "gmail_checked",
             "eml_auto_confirm",
@@ -316,8 +318,7 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
             "redmi_note",
             "accept_space", "accept_unsupported",
             "junk_hint",
-            "last_update_check", "last_announcement_check",
-            "notifications_reminder"
+            "last_update_check", "last_announcement_check"
     };
 
     @Override
@@ -456,6 +457,7 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
         swWebp = view.findViewById(R.id.swWebp);
         swAnimate = view.findViewById(R.id.swAnimate);
         swEasyCorrect = view.findViewById(R.id.swEasyCorrect);
+        swPastePlain = view.findViewById(R.id.swPastePlain);
         swInfra = view.findViewById(R.id.swInfra);
         swTldFlags = view.findViewById(R.id.swTldFlags);
         swJsonLd = view.findViewById(R.id.swJsonLd);
@@ -1537,6 +1539,13 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
             }
         });
 
+        swPastePlain.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+            @Override
+            public void onCheckedChanged(CompoundButton compoundButton, boolean checked) {
+                prefs.edit().putBoolean("paste_plain", checked).apply();
+            }
+        });
+
         swInfra.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton compoundButton, boolean checked) {
@@ -1697,7 +1706,7 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
         btnGC.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Helper.gc(true);
+                Helper.gc("Miscellaneous");
                 DB.shrinkMemory(v.getContext());
             }
         });
@@ -2369,6 +2378,7 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
             swWebp.setChecked(prefs.getBoolean("webp", true));
             swAnimate.setChecked(prefs.getBoolean("animate_images", true));
             swEasyCorrect.setChecked(prefs.getBoolean("easy_correct", false));
+            swPastePlain.setChecked(prefs.getBoolean("paste_plain", false));
             swInfra.setChecked(prefs.getBoolean("infra", false));
             swTldFlags.setChecked(prefs.getBoolean("tld_flags", false));
             swJsonLd.setChecked(prefs.getBoolean("json_ld", false));